requirements.txt

flask
sqlalchemy
pymysql
pandas
numpy
yfinance
ta
plotly
python-dotenv
xlrd
openpyxl